refactor directory name
